refactor sound
