+++ lib/game_runner.js
@@ -1,5 +1,6 @@
 const asteroids = require('../lib/asteroids');
 const Render = require('../lib/render');
+const Sound = require('../lib/sound');
 
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
@@ -9,19 +10,6 @@ var showLifeCount = false;
 function GameRunner(ship){
   this.ship = ship;
 }
-
-function Sound(){
-  this.sound = document.createElement("audio");
-  this.sound.src = 'assets/sounds/explosion_x.wav';
-  this.sound.setAttribute("preload", "auto");
-  this.sound.setAttribute("controls", "none");
-  this.sound.style.display = "none";
-  document.body.appendChild(this.sound);
-  this.play = function(){
-    this.sound.play();
-  }
-}
-
 
 var img = new Image();
 img.src = 'assets/images/nebula.png';
